# Performance improvements obtained by caching model-specific data representations and resamplings thereof