--- a/docs/make.jl
+++ b/docs/make.jl
@@ -26,5 +26,6 @@ makedocs(;
 ENV["TRAVIS_PULL_REQUEST"] = "false"
 
 deploydocs(;
-    repo="github.com/alan-turing-institute/MLJBase.jl.git",
+           repo="github.com/alan-turing-institute/MLJBase.jl.git",
+           push_preview=true
 )