--- a/src/machines.jl
+++ b/src/machines.jl
@@ -1,6 +1,6 @@
 ## MACHINE TYPE
 
-mutable struct Machine{M<:Model} <: MLJType
+mutable struct Machine{M<:Model,cache_data} <: MLJType
 
     model::M
     old_model::M # for remembering the model used in last call to `fit!`
@@ -11,7 +11,7 @@ mutable struct Machine{M<:Model} <: MLJType
     # `Source`s):
     args::Tuple{Vararg{AbstractNode}}
 
-    # model-specific reformatting of args:
+    # cached model-specific reformatting of args:
     data
 
     # cached subsample of data:
@@ -26,8 +26,9 @@ mutable struct Machine{M<:Model} <: MLJType
     # cleared by fit!(::Node) calls; put! by `fit_only!(machine, true)` calls:
     fit_okay::Channel{Bool}
 
-    function Machine(model::M, args::AbstractNode...) where M<:Model
-        mach = new{M}(model)
+    function Machine(model::M, args::AbstractNode...;
+                     cache=true) where M<:Model
+        mach = new{M,cache}(model)
         mach.frozen = false
         mach.state = 0
         mach.args = args
@@ -132,12 +133,14 @@ end
 
 
 """
-    machine(model, args...)
+    machine(model, args...; cache=true)
 
 Construct a `Machine` object binding a `model`, storing
 hyper-parameters of some machine learning algorithm, to some data,
 `args`. When building a learning network, `Node` objects can be
-substituted for concrete data.
+substituted for concrete data. Specify `cache=false` to prioritize
+memory managment over speed, and to guarantee data anonymity when
+serializing composite models.
 
     machine(Xs; oper1=node1, oper2=node2)
     machine(Xs, ys; oper1=node1, oper2=node2)
@@ -210,7 +213,7 @@ predictions = yhat(Xnew)
 """
 function machine end
 
-machine(T::Type{<:Model}, args...) =
+machine(T::Type{<:Model}, args...; kwargs...) =
     throw(ArgumentError("Model *type* provided where "*
                         "model *instance* expected. "))
 
@@ -219,33 +222,35 @@ static_error() =
                         "has no training arguments. "*
                         "Use `machine(model)`. "))
 
-function machine(model::Static, args...)
+function machine(model::Static, args...; kwargs...)
     isempty(args) || static_error()
-    return Machine(model)
+    return Machine(model; kwargs...)
 end
 
-function machine(model::Static, args::AbstractNode...)
+function machine(model::Static, args::AbstractNode...; kwargs...)
     isempty(args) || static_error()
-    return Machine(model)
+    return Machine(model; kwargs...)
 end
 
-machine(model::Model, raw_arg1, arg2::AbstractNode, args::AbstractNode...) =
+machine(model::Model, raw_arg1, arg2::AbstractNode, args::AbstractNode...;
+        kwargs...) =
     error("Mixing concrete data with `Node` training arguments "*
           "is not allowed. ")
 
-machine(model::Model, arg1::AbstractNode, arg2, args...) =
+machine(model::Model, arg1::AbstractNode, arg2, args...; kwargs...) =
     error("Mixing concrete data with `Node` training arguments "*
           "is not allowed. ")
 
-function machine(model::Model, raw_arg1, raw_args...)
+function machine(model::Model, raw_arg1, raw_args...; kwargs...)
     args = source.((raw_arg1, raw_args...))
     check(model, args...; full=true)
-    return Machine(model, args...)
+    return Machine(model, args...; kwargs...)
 end
 
-function machine(model::Model, arg1::AbstractNode, args::AbstractNode...)
+function machine(model::Model, arg1::AbstractNode, args::AbstractNode...;
+                 kwargs...)
     check(model, arg1, args...)
-    return Machine(model, arg1, args...)
+    return Machine(model, arg1, args...; kwargs...)
 end
 
 
@@ -335,6 +340,16 @@ function fitlog(mach, action::Symbol, verbosity)
     end
 end
 
+# for getting model specific representation of the row-restricted
+# training data from a machine, according to the value of the machine
+# type parameter `cache_data` (`true` or `false`):
+_resampled_data(mach::Machine{<:Model,true}, rows) = mach.resampled_data
+function _resampled_data(mach::Machine{<:Model,false}, rows)
+    raw_args = map(N -> N(), mach.args)
+    data = MMI.reformat(mach.model, raw_args...)
+    return selectrows(mach.model, rows, data...)
+end
+
 """
     MLJBase.fit_only!(mach::Machine; rows=nothing, verbosity=1, force=false)
 
@@ -392,7 +407,10 @@ and either `MLJBase.fit` (ab initio training) or `MLJBase.update`
 more on these lower-level training methods.
 
 """
-function fit_only!(mach::Machine; rows=nothing, verbosity=1, force=false)
+function fit_only!(mach::Machine{<:Model,cache_data};
+                   rows=nothing,
+                   verbosity=1,
+                   force=false) where cache_data
 
     if mach.frozen
         # no-op; do not increment `state`.
@@ -414,36 +432,36 @@ function fit_only!(mach::Machine; rows=nothing, verbosity=1, force=false)
     rows === nothing && (rows = (:))
     rows_is_new = !isdefined(mach, :old_rows) || rows != mach.old_rows
 
-    generating_new_resampled_data = rows_is_new && !(mach.model isa Static)
+    condition_iv = rows_is_new && !(mach.model isa Static)
 
-    data_is_valid =
-        isdefined(mach, :data)  &&
-        mach.old_upstream_state == upstream_state
+    upstream_has_changed = mach.old_upstream_state != upstream_state
 
-    # build or update `data` if necessary:
-    if !data_is_valid
-        generating_new_resampled_data = true
+    data_is_valid = isdefined(mach, :data) && !upstream_has_changed
+
+    # build or update cached `data` if necessary:
+    if cache_data && !data_is_valid
         raw_args = map(N -> N(), mach.args)
         mach.data = MMI.reformat(mach.model, raw_args...)
     end
 
-    # build or update `resampled_data` if necessary:
-    if !data_is_valid || rows_is_new
-        mach.resampled_data =
-            selectrows(mach.model, rows, mach.data...)
+    # build or update cached `resampled_data` if necessary:
+    if cache_data && (!data_is_valid || condition_iv)
+        mach.resampled_data = selectrows(mach.model, rows, mach.data...)
     end
 
-    # we `fit`, `update`, or return untouched:
-    if generating_new_resampled_data || force
+    # `fit`, `update`, or return untouched:
+    if mach.state == 0 ||       # condition (i)
+        force == true ||        # condition (ii)
+        upstream_has_changed || # condition (iii)
+        condition_iv
+
         # fit the model:
         fitlog(mach, :train, verbosity)
         mach.fitresult, mach.cache, mach.report =
             try
-                fit(mach.model, verbosity, mach.resampled_data...)
+                fit(mach.model, verbosity, _resampled_data(mach, rows)...)
             catch exception
-                @error "Problem fitting the machine $mach, "*
-                "possibly because an upstream node in a learning "*
-                "network is providing data of incompatible scitype. "
+                @error "Problem fitting the machine $mach. "
                 _sources = sources(glb(mach.args...))
                 length(_sources) > 2 ||
                     mach.model isa Composite ||
@@ -451,11 +469,18 @@ function fit_only!(mach::Machine; rows=nothing, verbosity=1, force=false)
                     @warn "Some learning network source nodes are empty. "
                 @info "Running type checks... "
                 raw_args = map(N -> N(), mach.args)
-                check(mach.model, source.(raw_args)... ; full=true) &&
+                if check(mach.model, source.(raw_args)... ; full=true)
                     @info "Type checks okay. "
+                else
+                @info "It seems an upstream node in a learning "*
+                    "network is providing data of incompatible scitype. See "*
+                    "above. "
+                end
                 rethrow()
             end
-    elseif mach.model != mach.old_model
+
+    elseif mach.model != mach.old_model # condition (v)
+
         # update the model:
         fitlog(mach, :update, verbosity)
         mach.fitresult, mach.cache, mach.report =
@@ -463,11 +488,14 @@ function fit_only!(mach::Machine; rows=nothing, verbosity=1, force=false)
                    verbosity,
                    mach.fitresult,
                    mach.cache,
-                   mach.resampled_data...)
+                   _resampled_data(mach, rows)...)
+
     else
+
         # don't fit the model and return without incrementing `state`:
         fitlog(mach, :skip, verbosity)
         return mach
+
     end
 
     # If we get to here it's because we have run `fit` or `update`!
@@ -713,16 +741,16 @@ function MMI.save(file::Union{String,IO},
 end
 
 # deserializing:
-function machine(file::Union{String,IO}, args...; kwargs...)
+function machine(file::Union{String,IO}, args...; cache=true, kwargs...)
     dict = JLSO.load(file)
     model = dict[:model]
     serializable_fitresult = dict[:fitresult]
     report = dict[:report]
     fitresult = restore(_filename(file), model, serializable_fitresult)
     if isempty(args)
-        mach = Machine(model)
+        mach = Machine(model, cache=cache)
     else
-        mach = machine(model, args...)
+        mach = machine(model, args..., cache=cache)
     end
     mach.state = 1
     mach.fitresult = fitresult