Depends on: https://github.com/alan-turing-institute/MLJModelInterface.jl/pull/76 which introduces `reformat` and `selectrows(::Model, I, args...)` to the MLJ model API. 

### Goals

Addresses some issues raised in #309, and in particular:

- Avoid the unnecessarily reconversion of data into a form specific to the model algorithm (eg, table -> matrix). Currently this happens every time model hyperparameter changes, or when training rows change.

- Allow resampling of data to happen at level of model-specific representation. Currently resampling is applied to the user-representation (eg, table). This would also allow implementation to use matrix views, for example,  avoiding copying, if this is appropriate. 

- Avoid unecessary resamplings of the data. Currently data is resampled on every `fit!` after a model hyperparameter change, even if the training rows are unchanged.

This PR is a proof of concept achieving these goals by:

- Adding new fields `data` and `resampled_data` to machines to cache model-specific representations of the data, and resampled versions respectively

- Change the logic in `fit_only!` to cache these representations, making use of the two methods `reformat` and  `selectrows(::Model)` in such a way that the fallbacks given in MLJModelInterface result in essentially the old behaviour. 

### Main features proposed solution

- **Buying into the improvements is optional** for existing (or new) implementations of the extended model API and so can be phased in gradually.

- Changes suggested here are only breaking to users who access current "low-level" methods `fit` and `update`, and only for those models that choose to buy into the performance improvements by implementing `reformat`, `selectrows(::Model)` and altering their `fit/predict` (see "Buying into the extended API" below). Users who interact using `fit!` should not encounter breaking behaviour. 

- Tested locally with the MLJModelInterface PR

- The current PR is limited to achieving the goals above on the training side, not on the predict/transform side

### Buying into the extended API

Here's an example of how to buy into the the new API, which is tested in detail [here](https://github.com/alan-turing-institute/MLJBase.jl/blob/b260019cd5f38ed7899b28109f003e3973cbac94/test/machines.jl#L176)

#### Existing implementation

```julia
function MLJModelInterface.fit(model::SomeModel,
                               verbosity,
                               X,
                               y)

    A = MLJModelInterface.matrix(X)
    fitresult = ...

    return fitresult, cache, report
end

function MLJModelInterface.predict(::SomeModel, fitresult, Xnew)
    Anew = MLJModelInterface.matrix(Xnew)
    return Anew*fitresult
end
```

#### Additions and changes

The signature of `fit` is changed to accept matrix `A` in place of table `X`, and the conversion to matrix is removed, with that role being played by the overloaded `reformat` method:
```julia

function MLJModelInterface.fit(model::SomeModel,
                               verbosity,
                               A,
                               y)
    fitresult = ...

    return fitresult, cache, report
end

MLJModelInterface.reformat(model::SomeModel, X, y) = (MLJModelInterface.matrix(X), y)
```

If `update` is also overloaded, then a similar change is made there. 

Finally, we must specify how to resample the model-specific representation of the data:

```julia
MLJModelInterface.selectrows(model::SomeModel, I, A, y) = (view(A, I, :), view(y, I))
```

That's it.