@yalwan-iqvia 2020-12-04
So cursory glance this seems good to me ( I didn't touch the code changes yet )

One of my main remarks would be that this seems geared towards implementors buying into the performance enhancements, however, I think a level of control needs to be afforded to the user on a case by case basis as to whether the implementation *should* buy in -- why? 

Well for some modelling/system setups, modelling might be getting near to memory limits in system. In this case caching resamplings can make the difference between running/not running.

In this case, I'd imagine user wishes to opt out in order to have it running with degraded performance, vs not running at all.

Does current implementation provide any mechanism for this?