@yalwan-iqvia 2020-12-04
P.s. Thanks for this

In LightGBM the system current has to construct LGBM specific data-sets for each resampling of data prior to fitting. This can actually be computationally non-trivial. By providing a mechanism by which we can hold ontoi our own form of the data and then producing our own resamples, this cost can almost entirely be negated. In fact, in my tests, the subsampling of the specific data structure is much faster than even just vanilla matrix slicing for large data sets (multi GBs +)