@ablaom 2020-12-04
@yalwan-iqvia Thanks for the feedback!

> In this case, I'd imagine user wishes to opt out in order to have it running with degraded performance, vs not running at all.

Maybe worth mentioning here that there are two kinds of caching - caching the model-specific representation of the whole training data set (before resampling) - and caching a particular subsample. Whether each adds to memory requirements depends both on the model implementation and what form the user-data takes. For example, if the "table" presented by the user is just a wrapped matrix (a `MatrixTable`) and the model algorithm is just unwrapping it, then there is no extra memory cost to "caching" the data. We are just pointing to stuff inside the user-specified data. Similarly, if subsampling of a matrix is just constructing a `view`, then again no extra memory cost. 

But, yes an "optimize_memory" option makes sense and can be added. However a particular model may want to also provide options - such as whether to construct a `view` of data or make a copy when restricting to specified rows of a matrix.

Can one construct a "view" of the model-specific representation in LightGBM (DMatrices?) (a view that can be consumed by `fit`). Or does indexing necessarily create copies? 

Be **great** if you could have a stab at implementing the extended API for LightGBM in a fork and do some benchmarking. The two branches you need - from MLJModelInterface - and from MLJBase for testing - are both called "mlj2". 

FYI I am on leave for the next week.