@yalwan-iqvia 2020-12-04
> Can one construct a "view" of the model-specific representation in LightGBM (DMatrices?) (a view that can be consumed by fit). Or does indexing necessarily create copies?

I *think* so

Basically you can create a Dataset object and then on top of this there is a getsubset API which returns a new dataset reference, but when I tested this with whomper data and then obtained a subset dataset I didnt see the memory usage change -- I believe (without looking at the C++ source code) that it just just a new dataset object which references the old one and additionally stores the subset elements request for use during training steps. Since this is still a dataset object it is consumable by lightgbm for subsequent fit calls.

We actually put a PR on the mainline code to allow this specialisation and performance improvement just recently:

https://github.com/IQVIA-ML/LightGBM.jl/pull/80/files

Idea was rather than requiring matrices, refactor out the dataset construction element and allow user to provide own dataset(s). This would then make it straightforward for us to support the paradigm within the MLJ wrapper (as the "user").

I can't promise I can get around to trying out the extended API over the next week but I'll certainly see what I can do