@jeremiedb 2020-12-09
Sorry being late to the show. Being able to circumvent the overhead was definitely a concern on my side, very nice to see an answer to it! I'll be testing the proposed buy-in from the `mlj2` branches with EvoTrees and hopefully there won't be any noticeable difference with the native API. I'll get back to you shortly. Let me know if there were any other aspect you'd like me to check.