@jeremiedb 2020-12-16
So the outcomes look positive from using the `mlj2` branches. 

For the model setup and initial iterations, the timing remains similar, but with a good reduction in the memory footprint: 

Old: 
``` julia
julia> @time fit!(tree, rows=train, verbosity=1, force=false)
[ Info: Training Machine{EvoTreeRegressor{Float32,…}} @230.
  3.089232 seconds (22.55 k allocations: 2.899 GiB, 18.94% gc time)
```

MLJ2:
```julia
julia> @time fit!(tree, rows=train, verbosity=1, force=false)
[ Info: Training Machine{EvoTreeRegressor{Float32,…}} @882.
  2.823589 seconds (21.94 k allocations: 1.701 GiB, 4.42% gc time)
```

Where things are interesting, is that while the backen `update` function was already performing well, the issue was coming from the call to `fit!` it situations of simple model update. On the new `mlj2` branches, that `fit!` function now performs pretty as fast as the underlying update. 

Old: 
```julia
tree.model.nrounds += 1
julia> @time update(tree.model, 0, tree.fitresult, tree.cache, X, Y);
  0.068055 seconds (1.94 k allocations: 41.474 MiB)

julia> @time fit!(tree, rows=train, verbosity=1)
[ Info: Updating Machine{EvoTreeRegressor{Float32,…}} @230.
  0.684355 seconds (2.78 k allocations: 658.003 MiB, 4.51% gc time)
```

New: 
```julia
tree.model.nrounds += 1
julia> @time update(tree.model, 0, tree.fitresult, tree.cache, X, Y);
  0.074739 seconds (1.95 k allocations: 41.473 MiB)

julia> @time fit!(tree, rows=train, verbosity=1)
[ Info: Updating Machine{EvoTreeRegressor{Float32,…}} @882.
  0.085758 seconds (2.22 k allocations: 41.507 MiB)
```

Which is good to see! 

Some small notes about implementation. 
To restrain the dependency to `MLJModelInterface`, the MLJBase.matrix is changed for: 
```julia
MLJModelInterface.reformat(model::EvoTypes, X, y) = (MLJModelInterface.matrix(X), y)
```` 

For the `metadata_model`, I've comment out `input=MLJModelInterface.Table(MLJModelInterface.Continuous)`.  What would be the proper way to specify the union of Table and Matrix?

Finally, this is a bit tangent, but has there been any adopted design for supporting the particular iterative training nature of boosted models? I'm referring to the difficulty from my perspective from using MLJ for such models given that the number of iterations isn't treated as other hyper-parameters, but rather more as an upper bound while the model is trained till an early stop criteria is met. It is over such sequence of iterations that one would want to perform an hyper search for GBT, but it doesn't appear trivial to integrate a tracking metric of an eval dataset in MLJ for that purpose, maybe I've overlooked this? That early stop over an eval data is something commonly embedded in GBT libraries, whether it's xgboost, lightgbm or evotrees.